requirements.txt
google-adk
python-dotenv
yfinance
duckduckgo-search
pandas-ta
pandas
numpy
uvicorn[standard]